Lesson 7: CommandParser works
