+++ TodoList/NoneCommand.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace TodoList;
+internal class NoneCommand : ICommand
+{
+	public void Execute()
+	{
+		Console.WriteLine("Неизвестная комманда.");
+	}
+}
