+++ TodoList/UncorrectCommand.cs
@@ -1,0 +1,9 @@
+﻿
+namespace TodoList;
+internal class UncorrectCommand : ICommand
+{
+	public void Execute()
+	{
+		Console.WriteLine("Не правильно введена комманда.");
+	}
+}
